requirements.txt
# Web framework
fastapi==0.111.0
uvicorn==0.30.1
uvloop==0.19.0; sys_platform != "win32"
httptools==0.6.1

# HTTP client
httpx==0.27.0

# Data validation
pydantic==2.7.1

# Resiliência
tenacity==8.2.3
pybreaker==1.2.0

# Observabilidade
prometheus-client==0.20.0
structlog==24.1.0

# OpenTelemetry (traces → Tempo)
# Ativado apenas se OTEL_EXPORTER_OTLP_ENDPOINT estiver definido
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-instrumentation-httpx==0.45b0

# Testes
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-cov==5.0.0
anyio[trio]==4.3.0

# Qualidade de código
ruff==0.4.4
bandit==1.7.8
